remove unnecessary test files
